requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.24.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dotenv==0.9.9
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.32.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html-to-markdown==1.9.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
lxml==6.0.1
microsoft-kiota-abstractions==1.9.6
microsoft-kiota-authentication-azure==1.9.6
microsoft-kiota-http==1.9.6
microsoft-kiota-serialization-form==1.9.6
microsoft-kiota-serialization-json==1.9.6
microsoft-kiota-serialization-multipart==1.9.6
microsoft-kiota-serialization-text==1.9.6
msal==1.33.0
msal-extensions==1.3.1
msgraph-core==1.3.5
msgraph-sdk==1.40.0
multidict==6.6.4
nh3==0.3.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
std-uritemplate==2.0.5
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
